requirements.txt
# ============================================
# FILE 5: requirements.txt (Updated with gdown)
# ============================================
# Flask & Server
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

# Core & Math
numpy==1.23.5
pandas==1.5.3
seaborn==0.13.2
joblib==1.5.1
tqdm==4.67.1
func-timeout==4.3.5

# Computer Vision & Image Processing
opencv-python-headless==4.9.0.80
scikit-image==0.24.0
lpips==0.1.4
ultralytics==8.3.174

# 3D Rendering & Geometry
chumpy==0.70

# Deep Learning Utils & Logging (CPU-optimized)
pytorch-lightning==2.5.5
tensorboard==2.8.0
torch-tb-profiler==0.4.3
protobuf==3.20.3

# Tools & Config
gdown==5.2.0
yacs==0.1.8
tabulate==0.9.0
termcolor==3.1.0
gitpython>=3.1.30

# Build & System
setuptools==70.0.0
cython==3.1.4
wheel==0.45.1
black==25.1.0
isort==6.0.1